Cleanup and add more splits
