+++ config/sly1.yaml
@@ -66,7 +66,7 @@ segments:
     - [0x1c008, asm, P2/splice/ref]
     - [0x1ca28, asm, P2/splice/serialize]
     - [0x1cd28, asm, P2/splice/sidebag]
-    - [0x1cf70, asm, P2/splice/spliceutils]
+    - [0x1cf70, cpp, P2/splice/spliceutils]
     - [0x1d128, asm, P2/splice/splotheap]
     - [0x1d508, asm, P2/splice/vecmat]
 
@@ -146,7 +146,7 @@ segments:
     - [0x5b8b0, asm, P2/flash]
     - [0x5bae0, asm, P2/fly]
     - [0x5d188, asm, P2/font]
-    - [0x5f810, asm, P2/frm]
+    - [0x5f810, c, P2/frm]
     - [0x60678, asm, P2/freeze]
     - [0x60f88, asm, P2/frzg]
     - [0x61070, c, P2/game]
@@ -203,7 +203,7 @@ segments:
     - [0x93068, asm, P2/pnt]
     #- [0x, asm, P2/po]
 
-    - [0x94210, asm, P2/prog]
+    - [0x94210, c, P2/prog]
 
     - [0x94598, asm] # TODO
     #- [0x, asm, P2/prompt]
@@ -211,9 +211,9 @@ segments:
     #- [0x, asm, P2/puffer]
     #- [0x, asm, P2/pzo]
     #- [0x, asm, P2/rail]
-    #- [0x, asm, P2/rat]
-    #- [0x, asm, P2/rchm]
 
+    - [0x9c038, asm, P2/rat]
+    - [0x9d440, asm, P2/rchm]
     - [0x9e410, asm, P2/render]
 
     - [0x9f040, asm, P2/rip]
@@ -260,9 +260,9 @@ segments:
     - [0xcd410, asm, P2/stephang]
     - [0xcefb8, asm, P2/stephide]
     - [0xd24f0, asm, P2/steppipe]
-    - [0xd39c8, asm, P2/steppower]
+    - [0xd39c8, c, P2/steppower]
 
-    - [0xd41d0, asm] #? unknown file
+    - [0xd41d0, asm, P2/step_pqr_unk] #? unknown file
 
     - [0xd46f0, asm, P2/steprun]
     - [0xd5148, asm, P2/stepside]
@@ -290,7 +290,7 @@ segments:
     - [0xeac78, asm, P2/update]
     - [0xeb408, c, P2/util]
 
-    - [0xec518, asm] #? unknown file
+    - [0xec518, asm, P2/uv_unk] #? unknown file
 
     - [0xee650, c, P2/vec]
     - [0xef318, asm, P2/vifs]
@@ -440,8 +440,11 @@ segments:
     #- [0x, data,, P2/ms]
     #- [0x, data,, P2/murray]
     #- [0x, data, P2/path]
-    #- [0x, data, P2/phasemem]
-    #- [0x, data, P2/pipe]
+
+    - [0x16aba4, data, P2/phasemem]
+    - [0x16aba8, data, P2/pipe]
+    - [0x16abac, data]
+
     #- [0x, data, P2/pnt]
     #- [0x, data,, P2/po]
     #- [0x, data, P2/prog]
@@ -489,7 +492,7 @@ segments:
     #- [0x, data, P2/stephide]
     #- [0x, data, P2/steppipe]
     #- [0x, data, P2/steppower]
-    #- [0x, data] #? unknown file
+    #- [0x, data, P2/step_pqr_unk] #? unknown file
     #- [0x, data, P2/steprun]
     #- [0x, data, P2/stepside]
     #- [0x, data, P2/stepzap]
@@ -509,7 +512,7 @@ segments:
     #- [0x, data,, P2/ui]
     #- [0x, data, P2/update]
     #- [0x, data, P2/util]
-    #- [0x, data] #? unknown file
+    #- [0x, data, P2/uv_unk] #? unknown file
     #- [0x, data, P2/vec]
     #- [0x, data, P2/vifs]
     #- [0x, data, P2/vis]
