+++ src/P2/frm.c
@@ -1,0 +1,29 @@
+#include "common.h"
+
+INCLUDE_ASM(const s32, "P2/frm", StartupFrame__Fv);
+
+INCLUDE_ASM(const s32, "P2/frm", OpenFrame__Fv);
+
+INCLUDE_ASM(const s32, "P2/frm", EnsureVu1Code__FP4VIFSPvT1);
+
+INCLUDE_ASM(const s32, "P2/frm", FinalizeFrameGifs__FP4GIFSPiPP2QW);
+
+INCLUDE_ASM(const s32, "P2/frm", FUN_0015ebf0);
+
+INCLUDE_ASM(const s32, "P2/frm", CloseFrame__Fv);
+
+INCLUDE_ASM(const s32, "P2/frm", PrepareGsForFrameRender__FP3FRM);
+
+INCLUDE_ASM(const s32, "P2/frm", check_anticrack_antigrab__Fv);
+
+INCLUDE_ASM(const s32, "P2/frm", FrameRenderLoop__FPv);
+
+INCLUDE_ASM(const s32, "P2/frm", RenderFrame__FP3FRMi);
+
+INCLUDE_ASM(const s32, "P2/frm", ClearPendingFrame__FP3FRM);
+
+INCLUDE_ASM(const s32, "P2/frm", FlushFrames__Fi);
+
+INCLUDE_ASM(const s32, "P2/frm", BlendPrevFrame__Fv);
+INCLUDE_ASM(const s32, "P2/frm", func_0015F618);
+INCLUDE_ASM(const s32, "P2/frm", func_0015F658);
