+++ src/P2/prog.c
@@ -1,0 +1,11 @@
+#include "common.h"
+
+INCLUDE_ASM(const s32, "P2/prog", __5CProgG4RGBAN31);
+
+INCLUDE_ASM(const s32, "P2/prog", Begin__5CProg);
+
+INCLUDE_ASM(const s32, "P2/prog", SetRemain__5CProgi);
+
+INCLUDE_ASM(const s32, "P2/prog", End__5CProg);
+
+INCLUDE_ASM(const s32, "P2/prog", Draw__5CProg);
