+++ src/P2/splice/spliceutils.cpp
@@ -1,0 +1,11 @@
+#include "common.h"
+
+INCLUDE_ASM(const s32, "P2/splice/spliceutils", StartupSplice__Fv);
+
+INCLUDE_ASM(const s32, "P2/splice/spliceutils", ShutdownSplice__Fv);
+
+INCLUDE_ASM(const s32, "P2/splice/spliceutils", spliceutils__static_initialization_and_destruction_0);
+
+INCLUDE_ASM(const s32, "P2/splice/spliceutils", _GLOBAL_$I$g_splotheapPair);
+
+INCLUDE_ASM(const s32, "P2/splice/spliceutils", _GLOBAL_$D$g_splotheapPair);
