+++ src/P2/steppower.c
@@ -1,0 +1,7 @@
+#include "common.h"
+
+INCLUDE_ASM(const s32, "P2/steppower", SetFsp__F3FSP);
+
+INCLUDE_ASM(const s32, "P2/steppower", UpdateJtActivePowerUp__FP2JTP3JOY);
+
+INCLUDE_ASM(const s32, "P2/steppower", IRotatePowerUp__FP3JOYiiPi);
